make sure get_delr and get_delc also work without extent in attributes
And change `assert ds.gridtype == "structured"` to `assert is_structured(ds)`
