+++ nlmod/dims/grid.py
@@ -36,63 +36,17 @@ from .layers import (
     remove_inactive_layers,
 )
 from .rdp import rdp
-from .shared import GridTypeDims, get_area, get_delc, get_delr  # noqa: F401
+from .shared import (
+    GridTypeDims,
+    get_area,
+    get_delc,
+    get_delr,
+    is_structured,
+    is_vertex,
+    is_layered,
+)  # noqa: F401
 
 logger = logging.getLogger(__name__)
-
-
-def is_structured(ds):
-    """Check if a dataset is structured.
-
-    Parameters
-    ----------
-    ds : xr.Dataset or xr.Dataarray
-        dataset or dataarray
-
-    Returns
-    -------
-    bool
-        True if the dataset is structured.
-    """
-    return GridTypeDims.parse_dims(ds) in (
-        GridTypeDims.STRUCTURED,
-        GridTypeDims.STRUCTURED_LAYERED,
-    )
-
-
-def is_vertex(ds):
-    """Check if a dataset is vertex.
-
-    Parameters
-    ----------
-    ds : xr.Dataset or xr.Dataarray
-        dataset or dataarray
-
-    Returns
-    -------
-    bool
-        True if the dataset is structured.
-    """
-    return GridTypeDims.parse_dims(ds) in (
-        GridTypeDims.VERTEX,
-        GridTypeDims.VERTEX_LAYERED,
-    )
-
-
-def is_layered(ds):
-    """Check if a dataset is layered.
-
-    Parameters
-    ----------
-    ds : xr.Dataset or xr.Dataarray
-        dataset or dataarray
-
-    Returns
-    -------
-    bool
-        True if the dataset is layered.
-    """
-    return "layer" in ds.dims
 
 
 def snap_extent(extent, delr, delc):
